devided code into more microservices and fixed bugs
