+++ src/extract/omdb_data_fetcher.py
@@ -3,7 +3,7 @@ import math
 import requests
 
 from src.extract.thread_pool_manager import ThreadPoolManager
-from src.config.config import Config
+from utils.config import Config
 
 
 class OMDBDataFetcher:
