+++ tests/test_transform.py
@@ -4,11 +4,11 @@ import os
 current_dir = os.path.dirname(os.path.abspath(__file__))
 sys.path.append(os.path.join(current_dir, '..'))
 
-from src.extract.extract import Extract
+from src.transform.transform import Transform
 from utils.config import Config
 
 if __name__ == "__main__":
     Config.validate_config()
 
-    e = Extract(5)
-    e.start()
+    t = Transform()
+    t.start()
